lfs: verify data when writing to indexedlog
Summary:
Currently we verify the content hash every time we read lfs data from indexedlog, but this is slow. Instead, we can verify once at insertion time and skip the verification when reading. This commit adds the insertion time verification.

Note that this introduces an error if the hash mismatches where previously the data would be inserted and silently ignored when reading.

Reviewed By: DurhamG

Differential Revision: D32444380

fbshipit-source-id: c096263f7a13279e21e31216a3a7132c52e630f2