cli_rs: Do not follow symlinks by using symlink_metadata
Summary: This was working fine for the usage counts we've aggregated so far, but when you try to aggregate the redirections, this gets stuck in infinite loops due to symlinks. Using `symlink_metadata` makes it not follow symlinks (symlink directories no longer pass the `is_dir()` check and are treated as a file). This matches what is done in the python version of `eden du`.

Reviewed By: fanzeyi

Differential Revision: D33160356

fbshipit-source-id: 81cdc638a262cad26aee29cbc93b1efbd26f23c5