Allow optimisation by reusing bubble
Summary:
This allows users to create a new snapshot reusing the latest snapshot's storage. This allows uploading very similar snapshots even faster (after the bugfix on D33096364).

There is already an optimisation to avoid re-uploading files. However, it still performs a get/put on the server side. This allows bypassing it altogether. The tradeoff is that the same bubble is used so we don't extend the lifetime.

In the future, we want to avoid the get/put on server side either way, but that needs a bunch more work.

Reviewed By: markbt

Differential Revision: D33098266

fbshipit-source-id: 94baad6d1db516a6300963d240c354c86a90fc05