Remove BlobRepo from log_commits_to_scribe
Summary:
**Context**: In addition to commit cloud commits, we want to log snapshots to Schism.

This is part of BlobRepo refactoring. It changes the `log_commits_to_scribe` function to use facet refs instead of BlobRepo, which changes a bit how it is used but maintains the behaviour. Some "helper" functions don't exist anymore, but the code actually looks cleaner.

This will be used later in the stack to use this function for snapshots, which don't have a BlobRepo.

Reviewed By: markbt

Differential Revision: D33100983

fbshipit-source-id: d3825ad3cd74c2b35d4d019e37f3f90f51118b5b