debugtop: add column for status of process
Summary:
This adds a column that shows whether a process is running. It shows `RUNNING` if the process is currently executing, or `TERMINATED (n)` if the process has finished. Here `n` is the exit code of the process (e.g. 0). Processes that terminate before debugtop starts running are not shown.

An option is added to the command for controlling the amount of time a terminated progress is shown after it finishes is also added.

Reviewed By: quark-zju

Differential Revision: D33522251

fbshipit-source-id: f8444298155aabecf4a33387b6ca56b67068367a