scsc: add case-insensitive option to scsc export
Summary:
Add the `--case-insensitive` option to `scsc export`.  If the given path
doesn't exactly find a file or directory, then `scsc` will look around for a
case-insensitive match for that same file or directory, and export the first
one it finds.

The search heuristics are that it starts to compare each path element, starting
at the end, case insensitively.   Once a match is found, it finds the first
case-insensitive match inside that subdirectory.  This should balance the
number of additional server requests for case insensitive look-ups with the
expectation that usually it will be the basename that differs in case (e.g.
`path/to/FileName.cpp` vs `path/to/filename.cpp`).

Reviewed By: farnz

Differential Revision: D33584150

fbshipit-source-id: c145ba28dc92da4020e0cb4024797b68c3cdb110