scsc: add case insensitive path-list-file handling
Summary:
If `--case-insensitive` is specified alongside `--path-list-file`, make the
application of the path-list-file case insensitive also.

The main caveat of this implementation is that we only take the *first*
case-insensitive match of any path element that we find.  In practice this
will mean the first path element in the real (case sensitive) ordering of
the manifest directory, but it should be considered to be an arbitrary one.

For files, this will have little effect on case-insensitive filesystems, as any
case-colliding files would alias anyway.

For directories, it means that if there are two directories with case-colliding
names then we will descend only into the first one that we find.
This should be sufficient for most uses, however if it is a problem,
we can adjust it later on.

Reviewed By: farnz

Differential Revision: D33584148

fbshipit-source-id: 657b69bdea7c2248b6b97d79101fda632bd33ae4