prjfs: handle notification in a single thread
Summary:
ProjectedFS is... weird. It can sometimes send notification out-of-order for
things like directory/file creation and removal and sometimes even send
duplicated removal notifications. This makes is really hard in EdenFS to keep
track of the changes in the inode hierarchy to behave similarly to NFS/FUSE
where the inode hierarchy always matches with what a user sees.

Previously, only out-of-order directory creation was handled correctly by
simply creating the parent directory if these weren't known by the inode
hierarchy, but this scheme fall over when handling directory removal. Consider
the following operations from the point of view of a user:

  mkdir a/b/c/d
  rmdir a/b/c/d
  mkdir a/b/c/d

EdenFS can receive the notifications in the following order:

  mkdir a/b/c/d
  mkdir a/b/c/d
  rmdir a/b/c/d

If EdenFS blindly follow the received notification, EdenFS will lose track of
the directory `d`.

Since deletion notification have been observed to also be sent multiple times
by ProjectedFS, EdenFS has no choice but to check the state of files on disk
and updating the inode hierarchy based on it. In order to avoid concurrency
issues, this is also done in a single thread.

For now, this is performed while servicing the notification and thus is
expected to slow down every writes, a future diff will return early from the
notification without waiting for the inode hierarchy to be updated to mitigate
this.

Reviewed By: chadaustin, fanzeyi

Differential Revision: D32329808

fbshipit-source-id: 33e01936fa49d2d8efa4ea3712f07dd2a132f3a4