always create a commit in small repos on mapping change
Summary:
Having a new commit on mapping change in every small repo is essential - the
commits based on old mapping can't be pushed using pushrebase which essentially
locks the repos for write. So far we've been lucky that the commit was created
because the directory with check-in mapping was bound to the small repo.

Now with two small repos the mapping-changing commit is empty in one of them so
it gets ommited. This commit changes this behaviour.

Reviewed By: farnz

Differential Revision: D33620294

fbshipit-source-id: 37cccb39131a978b27f838fc6b950fec83be9233