Use `select_biased` in derived data, instead of separate task
Summary:
`select_biased` has the same effect here as Abortable, but without spawning a second task. Use it, because it reduces the complexity

With Abortable, once we've flagged the derivation task as "should be aborted", we are stuck running it until it enters a `.await`, whereupon it returns.

With `select_biased`, we return to the macro every time the derivation task enters a `.await`, and we will preferentially check the first branch, which looks at the disabled watcher. If that watcher returns, we error out immediately, dropping the derivation task on the floor; if it enters the timer `.await`, we check the second branch (the derivation task).

Because dropping the unwanted timer is now done as part of finishing the derivation work, rather than in a separate task, we should see less contention for the timer wheel locks.

Reviewed By: mitrandir77

Differential Revision: D33622189

fbshipit-source-id: 6f7872535f117790669b450e29b079133581b2dc