mononoke: add --discovered-test arg to integration test runner
Summary:
The next diff adds granular target tests, this diff adds a feature to the runner so that they can successfully run.

When `buck test` is used, `tpx` first calls the integration_runner with the `--dry-run` flag. This allows the integration_runner to return a list of tests that are available. `tpx` will then iterate through that list, calling the `integration_runner` for each test and passing `run-tests,<test name>` once. This format is called the simple test selector.

With the previous logic, the test runner would return all ~300 tests during discovery and then tpx would call each test. Update the runner to accept multiple `--discovered-test` args, which the runner will return when `tpx` does test discovery. As these discovered tests aren't used during test running, the simple test selector that is provided is used.

In a later diff we can remove the existing test discovery and just rely on the tests that buck tells us about.

Reviewed By: HarveyHunt

Differential Revision: D33630573

fbshipit-source-id: 20dcd0578ed6e193f8459bdc7acb8820cee7f3b0