cli: improve error message when cloning over an existing clone
Summary:
When EdenFS is stopped, users have attempted to remove the root directory of
the mount point, in the hope that it deletes their repository. Cloning a new
repository at that location gives them an unhelpful error message claiming that
the directory isn't empty:

  error: [Errno 41] Directory not empty: 'C:\open\foobar'

Let's improve this error to help guide the user towards using `eden rm`.

Reviewed By: kmancini

Differential Revision: D33574906

fbshipit-source-id: 3f2ae567dd9b2af2493e6e52c52c85a8216c993b