git: provide a higher level pull
Summary:
Previously, `pull` takes refspecs. The refspec is a git detail. It forces the
callsite to understand the git refs and git-fetch refspec. Provide a higher
level `pull` so the callsite only needs to understand remote names (ex. "main")
and commit hashes.

This makes `git.pull` more similar to `repo.pull`, and makes it easier to add
auto pull and `hg paths` features to git repos.

Reviewed By: DurhamG

Differential Revision: D33518825

fbshipit-source-id: ed109b42e26c897c7f9569c00f6630709372a258