prjfs: handle materialized directories with no backing store
Summary:
When an update removes a directory but a materialized ignored file/directory
was present, EdenFS will rightfully not remove that ignored file/directory.
However, the parent directory would have been a placeholder on disk. And since
placeholder directories can never be made full[0], the directory keep its
placeholder. Listing this directory would thus issue an opendir callback, but
EdenFS would fail due to that directory not being present in the manifest.

Since this case can only happen in this case, let's simply return an empty
directory and let ProjectedFS merge it with the on-disk directory content.

[0]: https://docs.microsoft.com/en-us/windows/win32/projfs/cache-state

Reviewed By: genevievehelsel

Differential Revision: D33676621

fbshipit-source-id: bdbf769175a6471a1a709bdb0b03a45bd60ef366