inodes: release EdenMount parentLock during checkout
Summary:
On Windows, the parentLock needs to be acquired on PrjFS callbacks in order to
walk the Tree hierarchy. Unfortunately, since checkout holds this lock for the
duration of checkout operation, the callbacks will be blocked. This is usually
not an issue as update is expected to be relatively fast, but in some cases, an
update may trigger these callbacks. One of these can happen when EdenFS is
killed midway through a checkout operation and checkout tries to resolve
conflicts during a force checkout, the current code will iterate through many
unloaded inodes and issue recursive callbacks.

While the above issue can be fixed, we can also rethink the reason for holding
the lock for the duration of checkout. The main one being to prevent concurrent
checkout, and this can be achieved by storing a boolean protected by the lock.
That way we can detect the pending checkout, and not hold the lock for too
long.

Reviewed By: chadaustin

Differential Revision: D33668308

fbshipit-source-id: 634c60d4684c755d38e4b26826e3b14afbc6083a