allow integration tests that can not run on NFS
Summary:
In one of my next changes I am going to be introducing a integration test where
NFS is not enabled.

Previously, we did not have any tests that ran with NFS disabled and so a bug
slipped through the cracks in graceful restart when NFS is disabled.

If I am going to disable NFS then these tests should not pretend to run on NFS.
This introduces a parameter to the eden test decorator that allows skipping
the NFS tests.

Reviewed By: genevievehelsel

Differential Revision: D33247961

fbshipit-source-id: ab37bd1aa64b9b09a69fe5e90e8068d4f444cd29