Back out "inodes: speedup overlay initialization on Windows by 15x"
Summary:
We've started seeing hangs during update on Windows due to EdenFS trying to
invalidate files and directories that do not have a placeholder already. When
that happens, ProjectedFS will send a lookup callbacks to EdenFS to write a
placeholder to disk. This triggers a recursive callback to EdenFS, which
unfortunately ends up deadlocking EdenFS.

The reason EdenFS is trying to invalidate these files and directories, is due
to the original diff eagerly loading more inodes to speedup startup time. Since
this original diff is causing so much issues, let's revert it.

Reviewed By: kmancini

Differential Revision: D33697483

fbshipit-source-id: 6c15c9cfbd784e1033c2096e1794929b888bbbd7