revert: don't backup symlinks
Summary:
When doing a revert, Mercurial keeps a copy of reverted files in the
.hg/origbackups directory. If we put symlinks in this directory it can mess up
future file backups, since they might accidentally be put inside the symlink
(which can result in a failure if the symlink is to another device).

Since these backups are almost never used, and are relatively janky, let's just
never backup symlinks.

Differential Revision: D33689167

fbshipit-source-id: 0b227b84122c84e7ea1d12e5219e6b6946c882af