use direct mononoke connection in test-cmd-segmented-changelog-tailer-race
Summary:
The graph I deleted has a diamond in it and it can be rendered differently based on which commit was processed by segmented changelog earlier.
There was a difference between OSS and regular test environments.

```
@@ -167,9 +167,9 @@
   ├─╯
   o    5c99bc570462 X: merge branch 1
   ├─╮
-  │ o  cd40fe85b470 X: commit 1 for branch 1
+  │ o  49c45f24d44a X: base branch 1
   │ │
-  o │  49c45f24d44a X: base branch 1
+  o │  cd40fe85b470 X: commit 1 for branch 1
   ├─╯
   o  26805aba1e60 C
   │
```

The problematic part of output is not necessarily needed. We were printing out the graph that that was just imported to mononoke.
We print it out locally, save it and then later compare in the line with `diff` at the beginning to the graph from mononoke.

Reviewed By: HarveyHunt

Differential Revision: D33706427

fbshipit-source-id: 925219f347816c6cef4e0ded017aa1ac9302383e