cli_rs: Print out failed file checks
Summary:
When aggregating disk usage size, we may fail to check some file sizes. Let's print those out.

One notable difference is that I changed all of them to be HashSets in case there was some overlap between paths. I'm not sure if this would happen or not -- especially for the mount failed files case.

Reviewed By: fanzeyi

Differential Revision: D33242908

fbshipit-source-id: eb32a24d4b5bdd4e49bc49c282a7ce95925347fa