mononoke_app: add arg extensions
Summary:
Add `ArgExtension`, a trait that can be implemented to extend the arguments of
a Mononoke app, and insert processing of these arguments into the start-up
processing in order to modify the environment before it is used to start
Mononoke.

Using a callback model allows us to avoid additional dependencies for
`mononoke_app` when adding custom arguments to individual binaries, e.g. scrub
arguments, or mcrouter arguments.

Reviewed By: mitrandir77

Differential Revision: D33741137

fbshipit-source-id: b63fd8a209b0c0d12fd1260753384868c69ab067