mononoke_app: add support for computed defaults
Summary:
Add support for computed defaults of values.  These can either be set
by the caller, using `with_default` or `with_defaults`, or built from default
values in the builder.

Use this initially for readonly storage, so that the shown default matches what
the app builder sets.

To make the lifetimes work, we need to change the builder to work by
reference, so that the compiler knows that the builder (which contains
the defaults) will outlive the subcommands that we will add to the app.

Reviewed By: mitrandir77

Differential Revision: D33719647

fbshipit-source-id: 731e931aed4a5f7b7ed5c1850bab6e68362f3dd3