prjfs: use shared_ptr for ObjectFetchContext
Summary:
In the case of notifications, these are running in a background thread,
potentially after the original context has been destroyed. This would therefore
lead to use after free. By using shared_ptr, the lifetimes are easier to
understand and less error prone, so let's use them in the dispatcher.

Reviewed By: chadaustin

Differential Revision: D33724774

fbshipit-source-id: 048a62a66e9ef09369cc66f9547501cf7316863f