inodes: remove use of TreeInode in Prjfs lookup callback
Summary:
ProjectedFS callbacks may be caused by EdenFS itself in some situations. One of
them being reading of materialized files to compute their sha1, another one
being invalidation during checkout if EdenFS has inodes being loaded that aren't
placeholder on disk.

In that second case, several inodes locks are held, including the RenameLock
and the TreeInode content lock for the directory being invalidated. For this
reason, trying to access one of these locks in a ProjectedFS callback can lead
to a deadlock.

The lookup callback unfortunately may try to access these locks, and while in
normal running condition EdenFS shouldn't be invalidating files/directories
that aren't placeholder on disk, we've seen cases where it can happen.

The reason this callback tries to access these locks is due to EdenFS relying
on the Inode refcount to decide if an Inode needs to be invalidated during
checkout. Without the refcount increment, EdenFS would simply not invalidate
it, and thus the wrong content may be present on disk after checkout. However,
this increment doesn't need to happen in the same thread, and it can be pushed
to the same executor that is used to handle write operations in EdenFS. The
main reason being that notifications on files can only occur after a file is
made a placeholder, and thus we can guarantee that the refcount will always be
incremented prior to that file being modified. The executor also guarantees
that checkout will wait for all pending notifications before starting, and thus
it guarantees that the refcount will always be incremented prior to checkout
starting.

Reviewed By: chadaustin

Differential Revision: D33724773

fbshipit-source-id: 7db293363da181a6d6a70c78d6930c5320694b41