print testtmp dir before test run only if keep-tmpdir passed
Summary:
This is especially useful when there is long-running test or a test that's stuck. With this path it will be possible to at least look at logs or configs that are being used in the test environment.

I also considered outputting it by default if isatty() == True, but some folks might not like that and it might be too noisy.

Reviewed By: yancouto

Differential Revision: D33792360

fbshipit-source-id: 5c440c3be63242791d92b0decb523c70f4527ee9