inodes: move prjfs directory invalidation to a different thread
Summary:
Directory[0] invalidations can trigger recursive ProjectedFS callbacks onto
EdenFS, when this happens we need to make sure that the calling context doesn't
hold any locks/threads that the callback will need to run. We otherwise can run
into a deadlock situation where the callback may try to acquire a lock held by
the calling context.

In this situation, checkout is this calling context, and it has been shown that
directory invalidation may in some cases trigger a recursive lookup. While the
lookup callback doesn't use any shared locks, it pushes work to the server
state thread pool in `ObjectStore::getTree`. Unfortunately, this thread pool is
also where checkout runs, and thus if too many directory invalidation are
occuring at the same time, all the threads in the thread pool might be blocked
and thus `ObjectStore::getTree` may never complete, causing the lookup callback
to not complete, causing the invalidation to not complete, causing checkout to
not complete.

The solution is to move the invalidation away from the thread pool, that way
checkout doesn't block in the thread pool while the invalidation is ongoing,
allowing `ObjectStore::getTree` to push work onto it and run it.

[0]: File invalidation won't trigger these recursive callbacks.

Reviewed By: chadaustin

Differential Revision: D33773892

fbshipit-source-id: a6516afdc5c68ac8d1d7ba75f91c45fb12cc6ee4