Test for segmented changelog pull
Summary:
This is part of solving T110433574.

Ideally D32800580 (https://github.com/facebookexperimental/eden/commit/e704fd5605370ea21f0ed48618e9fbb907513e23) would've been caught by an already existing test, as it broke basically all pulls, and we already have tests that pull.

Unfortunately, our tests often don't use the same configs as prod, in this case segmented changelog.

On this diff, I add a test that catches the bug (in fact, a fix needs to be landed before this diff), by doing some e2e push/pulls using the most modern stuff on the client/server (edenapi stuff, and segmented changelog).

It was quite harder than expected to get this working, there are **a lot** of config options needed that aren't on by default on tests, but are rolled out 100% (or almost) to prod. This is a good test to have as it catches the issue, and can be used as an example for future tests for how to set up config (as it's tricky).

Tomorrow, on following diffs, I'll try to move as many of these options as possible to be **default for all tests**, and hopefully I'll be able to surface the errors on D32800580 (https://github.com/facebookexperimental/eden/commit/e704fd5605370ea21f0ed48618e9fbb907513e23) on other already existing tests, as they should've been, or at least get closer to that and increase the coverage of the codepaths that actually matter in production.

Reviewed By: quark-zju

Differential Revision: D33749312

fbshipit-source-id: 5744c4ff89ea166bcdc5c8ef56f31ccf2f7162c4