Extract retry function from hg_sync_job to common lib
Summary:
This diff extracts the `retry` function to a more common place. This is used to retry arbitrary functions, and from a quick search you can find multiple implementations of this throughout mononoke ([1]() [2]() [3]()), with different level of features (logging, exponential backoffs, splaying).

I made it slightly more generic, by using `FnMut` and templating the error type, so it's more future proof.

This seems to be the most generic one, so I'm extracting it to reuse it in a future diff, and avoid adding yet another implementation.

Reviewed By: ahornby

Differential Revision: D33791605

fbshipit-source-id: 6a09093a3334d41491ecab779a90b23b459d7db3