remove hgcli code path from mononoke server
Summary:
No hgcli requests: https://fburl.com/ods/vkr7tjs7

Hgcli was a small binary that was running on hg servers. Its purpose was to proxy users' ssh traffic landing on hg servers to mononoke. It was sending a custom Preamble to mononoke with users' info parsed on hg servers from ssh environment, and after that it was proxying wireproto traffic that mononoke understands.

Now no client connects to hg servers. Everyone connects to mononoke using http (either http with upgrade to wireproto (only push) or edenapi). Hgcli is not even installed on hg servers. There is on need to have that code path around.

Reviewed By: yancouto

Differential Revision: D33841713

fbshipit-source-id: 682594210789ba5669660c53634f2f8111f39e41