sparse: accept relative sparse profile names
Summary: Make "sparse enable" normalize profile names. It is not clear it wants a canonical path from the root of the repo, so now it makes some attempts to convert the given path(s) to that form. It now works with paths relative from cwd, and normalizes the path separator.

Reviewed By: quark-zju

Differential Revision: D33851998

fbshipit-source-id: 91a3d59b91266bb8941408d23d5021fa1f9e7025