tools/admin: implement blobstore subcommand
Summary:
Implement the `blobstore` subcommand for examining and uploading to the
blobs in the blobstore.

This is an improvement over the existing `blobstore-fetch`, as we can now
interpret the data as more types, and hex-dump the data if we can't decode it
as a known type.

Not all types are implemented yet, but it should be easy to add them later.

Reviewed By: mitrandir77

Differential Revision: D33791723

fbshipit-source-id: bb2ea6c76986fcdd34b3d7be773c5c8168e41be8