service: fix use after free in getAttributesFromFiles
Summary:
When write notifications are pending on Windows, waitForPendingNotification
would wait for some time. Unfortunately, this means that the mountPoint thrift
argument would get freed, and since the lambda only captured a piece of the
path, and not the path directly, the mount path would get used after the
underlying path has been freed. We need to copy the path to avoid this issue.

Reviewed By: MichaelCuevas

Differential Revision: D33927317

fbshipit-source-id: 790bcb26cb46cdac750826ebe13c56e0d1275c22