config: update SNAPSHOT with in progress checkout information
Summary:
If a checkout is interrupted midway, EdenFS is unable to recover without
fetching very large parts of the repository. This leads to a poor user
experience as Mercurial suddenly hangs when doing anything.

In order to solve this, let's introduce a new SNAPSHOT version that allows
tracking the ongoing checkout operation and the commit origin and destination.

Since there is no current way for EdenFS to recover from this, let's also ask
the user to reclone if EdenFS is killed/crashed during checkout.

Reviewed By: genevievehelsel

Differential Revision: D33859121

fbshipit-source-id: b7cc4d2bc7aade047a230f5b579d8461c8586451