Make edenapi-e2e test less flaky
Summary:
This is already done in another `hg pull` in this test, but needed to be done in the clone as well, as the test is a bit flaky right now.

Ideally we need to fix this so we don't need to sleep anymore, but this will unblock us for now. I tried flushing WBC in another diff (D33771281) but it did not fix the whole issue, there might be more underneath.

Reviewed By: markbt

Differential Revision: D33917157

fbshipit-source-id: c78b3718cb4d32480d1dce8dc2c962928f3f79ad