Default pull.httpmutations to true
Summary:
## Context

There are some config options that have been rolled out to 100% of users for some time, but they're still not used in tests by default, which makes our coverage worse.

The best way to change the behaviour of hg regarding config is to change the default value on `configitems.py`, and then (eventually, when this rolls out everywhere), delete them from the configerator overrides config.

The other option would be to leave them in configerator (or in eden/scm/lib/configparser/src/fb/mod.rs), but then they would also need to be replicated in all tests that want to use it, or at the very least duplicated in the base hgrc for integration tests, which is very prone to errors (or simply forgetting to do it, the state we are in today).

Notice that this still allows us to set them to false via configerator quickly, if necessary.

## This diff

This diff defaults `pull.httpmutations` to true: It was previously always set to true on configparser/src/fb/mod.rs

Reviewed By: quark-zju

Differential Revision: D33771940

fbshipit-source-id: 987db9cfaff720645eac4ad65edf560f875b3655