teach `eden doctor` to recognize stale NFS mounts
Summary:
`eden doctor` does not know how to detect or fix broken NFS mounts.

Let's teach it to recognize NFS mounts, recognize when those mounts are stale,
and fix it.

Recognizing a stale mount is a bit different on NFS and FUSE. FUSE is kind and
responds with errors when you try to read a stale mount. NFS is not so kind.
It just hangs. So instead of certain errors indicating stale mounts, hangs
will have to indicate them.

Reviewed By: chadaustin

Differential Revision: D33283361

fbshipit-source-id: 939c212db1dcb1904fc8bbf3bbf32cd39970a101