Special-case linkfile src="."
Summary:
I am making it impossible to put "." or ".." as a path element in Mononoke's `MPath` abstraction, since Mononoke does not do path traversal, and canonicalization is a minefield.

However, `linkfile src="."` occurs in https://fburl.com/code/evq63pto as a special case. Treat it specially in our code, so that I can forbid path traversal type elements in an `MPath`.

Reviewed By: mojsarn

Differential Revision: D33952767

fbshipit-source-id: 0c4ce68014fb6877163035d1c7eb813616748dd2