repo: add method for repo initialization
Summary:
Adds a method that will be used in the future for replacing the repo initialization done in Python.

More specifically, this method creates and populates the files required for a new repo in the same way as we currently do via the constructor of `localrepo` with the extensions and configs we have enabled by default.

Reviewed By: DurhamG

Differential Revision: D33901712

fbshipit-source-id: 0d12732f5c9d36fb115a97e53b356a88fd609b6c