lfs: retry fetch requests on 503 response code
Summary:
Currently a single LFS 503 during clone/checkout aborts the operation and leaves the user in a situation that is difficult to recover from. Now we treat a 503 as any other 5XX server error and retry.

The 5XX retry count/backoff can be configured via the lfs.backofftimes config knob. It defaults in the code to [1, 4, 8], which is 3 retries with rand(1), rand(4) and rand(8) sleep in between.

503 can be served from proxygen indicating there are no live lfs servers, or from x2pagentd in cases of network failure.

Reviewed By: quark-zju

Differential Revision: D33804601

fbshipit-source-id: cbb3514051843d3517bb676cf7c83dce057e42b9