inodes: speedup overlay initialization on Windows by 15x take 2
Summary:
The original had to be reverted due to causing checkout to recursively call
into EdenFS which was making EdenFS deadlock. These issues have now been solved
and even if EdenFS were to recursively cause a callback to be triggered, this
wouldn't lead to a deadlock.

To make sure this is true, let's have the fast code behind a config so we can
roll it slowly to user and roll it back immediately if needed.

Reviewed By: genevievehelsel

Differential Revision: D33950740

fbshipit-source-id: 2a91e4aabc7466bb710ec7c65d3c9a10c972842b