telemetry: make RequestMetricsScope::getMaxDuration constant time
Summary:
Since watches are always added at the end of the list, and time is
monotonically increasing, requests are implicitely ordered by their running
time in the list, with the longest running one at the front. Thus in order to
compute the max duration, we can simply look at the front of the list.

Reviewed By: kmancini

Differential Revision: D33994895

fbshipit-source-id: 3c4f10da60b7554126a08aad7042cc4c712699c3