init: integrate Rust repo initialization into the `init` command
Summary:
This makes use of the new Rust repo initialization added in previous commits into the `init` command in Python by replacing the initialization that was done mostly by the `localrepo` object. The codepaths for `git` and `bundle` remain active.

Another smaller addition was the option of removing `narrowheads` from the list of requirements if `experimental.narrow-heads` was explicitly disabled in the configurations.

Reviewed By: quark-zju

Differential Revision: D33901725

fbshipit-source-id: d9d976a1b844e1dd6a5e065f8ec8bf1d04fc4c03