dag: define IdDagStore::{resize,remove}_flat_segment
Summary:
See D33779403 for context. This is a step forward to implement "strip".
This diff defines the API that can be used to remove or "shrink"
a flat segment and adds some sanity checks.

The actual implementation is in upcoming diffs.
Testing will focus on "remove" first, "resize" tests will be added later.

Differential Revision: D33821386

fbshipit-source-id: 61848593e698fe7f9358b25a4cb72788637c50a5