ignore file not found when creating a snapshot
Summary:
Sometimes hg might recognise an untracked file as being added, but by the time it tries to load it, the file is gone, causing a crash.

For snapshots, it should be fine in this case to simply ignore this file. That is what this diff does. It specifically ignores `FileNotFoundError`, meaning other files still get classified as errors and we're not swallowing anything we shouldn't be.

Reviewed By: mitrandir77

Differential Revision: D34043083

fbshipit-source-id: cec7d528652c1704f3a3db707eafc0e43f24ac74