Add context to failure to create snapshots
Summary:
This just adds a `with_context` call to when we're loading files from disk, so it prints which file failed to load.

I've seen several `[Errno 2] No such file or directory (os error 2)` and they're really annoying.

To do this, I had to refactor the lambda into a function, because try blocks are not stable yet :(.

Reviewed By: markbt

Differential Revision: D34043084

fbshipit-source-id: e0fd4142d75d6b392440f7763217a9b23ee8c8a4