vfs: chmod lock files to 0666 on unix
Summary: Unlike the python repo lock files, the rust lock files are not deleted when the lock is released. That means if the locks are created as root, they will break future commands run as other users. Try to mitigate that eventuality by chmoding the lock files to 0o666 on unix.

Reviewed By: quark-zju

Differential Revision: D34016719

fbshipit-source-id: 2bd1ec9d5da0e27a12eb95da5c2a4cb3f6286137