reduce verbose logging
Summary:
This case is getting hit quite a lot on NFS mounts on M1. This is filling up
the logs.

We need to investigate why this is getting hit so often since its a bit unexpected
how often this case is hit, but to reduce the load of the logs for now we can
turn off this log in production.

Reviewed By: fanzeyi

Differential Revision: D34112646

fbshipit-source-id: 96cc8e096c0cc3388f14ff2e667565f3bc3911b8